voting user and product

+++ src/main/java/com/salespage/salespageservice/app/controllers/PrivateProductController.java
@@ -213,7 +213,7 @@ public class PrivateProductController extends BaseController {
     }
 
     @PostMapping("rating")
-    public ResponseEntity<BaseResponse> updateRating(Authentication authentication, @RequestParam String productId, @RequestParam Long point) {
+    public ResponseEntity<BaseResponse> updateRating(Authentication authentication, @RequestParam String productId, @RequestParam Float point) {
         try {
             return successApi("Đánh giá sản phẩm thành công", productService.updateRating(getUsername(authentication), productId, point));
         } catch (Exception ex) {
